pyproject.toml
[tool.poetry]
name = "idealista"
version = "0.1.0"
description = ""
authors = ["angelpuertadiaz@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dacite = "^1.8.0"
toml = "^0.10.0"
requests = "^2.28.2"
google-api-python-client = "^2.131.0"
pandas = "^2.2.2"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
